refactor: 지역 관련 상수 분리
